requirements.txt
asgiref==3.5.0
autopep8==1.6.0
Django==4.0.1
django-rest-framework==0.1.0
djangorestframework==3.13.1
Pillow==9.1.0
djangorestframework-simplejwt==5.1.0
drf-extensions==0.7.1
drf-nested-routers==0.93.4
factory-boy==3.2.1
Faker==12.3.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2